+++ srai/loaders/download.py
@@ -3,15 +3,10 @@
 import warnings
 from pathlib import Path
 
-import requests
-from tqdm import tqdm
-
-from srai.constants import FORCE_TERMINAL
+from pooch import retrieve
 
 
-def download_file(
-    url: str, fname: str, chunk_size: int = 1024, force_download: bool = True
-) -> None:
+def download_file(url: str, fname: str, force_download: bool = True) -> None:
     """
     Download a file with progress bar.
 
@@ -23,29 +18,21 @@ def download_file(
 
     Source: https://gist.github.com/yanqd0/c13ed29e29432e3cf3e7c38467f42f51
     """
-    if Path(fname).exists() and not force_download:
+    destination_path = Path(fname)
+    if destination_path.exists() and not force_download:
         warnings.warn("File exists. Skipping download.", stacklevel=1)
         return
 
-    Path(fname).parent.mkdir(parents=True, exist_ok=True)
-    resp = requests.get(
-        url,
-        headers={"User-Agent": "SRAI Python package (https://github.com/kraina-ai/srai)"},
-        stream=True,
-    )
-    resp.raise_for_status()
-    total = int(resp.headers.get("content-length", 0))
-    with (
-        open(fname, "wb") as file,
-        tqdm(
-            desc=fname.split("/")[-1],
-            total=total,
-            unit="iB",
-            unit_scale=True,
-            unit_divisor=1024,
-            disable=FORCE_TERMINAL,
-        ) as bar,
-    ):
-        for data in resp.iter_content(chunk_size=chunk_size):
-            size = file.write(data)
-            bar.update(size)
+    destination_path.parent.mkdir(parents=True, exist_ok=True)
+    if force_download:
+        destination_path.unlink(missing_ok=True)
+
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore")
+        retrieve(
+            url,
+            fname=destination_path.name,
+            path=destination_path.parent,
+            progressbar=True,
+            known_hash=None,
+        )