+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped `QuackOSM` version to `0.14.0`
 - Bumped `OvertureMaestro` version to `0.4.2`
+- Refactored download function to use `pooch` library instead of `requests`
 
 ## [0.9.7] - 2025-05-21